requirements.txt
Django==4.2.7
Pillow==10.0.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-allauth==0.57.0
python-decouple==3.8
django-extensions==3.2.3
python-telegram-bot==20.7
